Remove additional files that should not be packed inside the course archive
The list of additional files inside `course-info.yaml` lists the files necessary for the learners. These files will be packed to the course archive delivered to learners. Normally, the JetBrains Academy plugin correctly decides which files to add as additional, but sometimes it mistakenly adds or skips some files. In that case, the course author should add or remove them manually.

The unneeded files from /cmake-build-debug/ are removed from the list. The file .courseignore (added in the previous commit) helps the plugin to skip /cmake-build-debug/* files in the future.